backend/requirements.txt
Django>=4.0
psycopg2-binary
djangorestframework
python-dotenv
django-cors-headers
dj-database-url
djangorestframework-simplejwt

